[FIX] Internal audit: migrate remaining Integer/Bool sites in attrs/configs/registries
Internal audit on the partial-landing of PR #19614 found several
attr/registry sites that were missed. Migrated them following the
same Integer/Bool -> int64_t/bool pattern.

OpAttrMap registries (TCallEffectKind, TScriptDtypePrintLocation)
----------------------------------------------------------------
The two operator-attr-map type aliases in
`include/tvm/tirx/op_attr_types.h`:

    using TScriptDtypePrintLocation = Integer;  -> int64_t
    using TCallEffectKind            = Integer;  -> int64_t

Updated every `set_attr<TCallEffectKind>(..., Integer(K))` and
`set_attr<TScriptDtypePrintLocation>(..., Integer(K))` writer in
`src/target/{cuda,metal,webgpu}/intrin_rule_*.cc` and
`src/tirx/op/{op,builtin,runtime,target_builtin/*}.cc` and
`src/tirx/ir/stmt.cc` to use `static_cast<int64_t>(K)`.

Updated the 4 readers (`Op::GetAttrMap<TCallEffectKind>(...)[op]`
and the `TScriptDtypePrintLocation` reader in
`src/tirx/script/printer/expr.cc`) to drop `->value` /
`.IntValue()` and explicitly cast the int64_t back to the enum.

Function-attr / DictAttrs / WithAttr / attr_map.Set
----------------------------------------------------
Replaced `tvm::Bool(true)` with bare `true` at:

  - `src/te/operation/create_primfunc.cc` (2x DictAttrs init)
  - `src/relax/transform/fuse_tir.cc` (attr_map.Set)
  - `src/relax/transform/compute_prim_value.cc` (DictAttrs init)
  - `src/tirx/script/builder/frame.cc` (insert_attr for kSTir and
    kPersistentKernel)
  - `src/relax/backend/vm/codegen_vm_tir.cc` (WithAttr)
  - `src/relax/backend/vm/vm_shape_lower.cc` (WithAttr)
  - `src/tirx/transform/split_host_device.cc` (WithAttr)

Replaced `WithAttr(..., "scoped", Integer(1))` with
`WithAttr(..., "scoped", static_cast<int64_t>(1))` in
`src/relax/transform/specialize_primfunc_based_on_callsite.cc`.

Buffer/loop/block annotations
------------------------------
`annotations.Set(kVolatile, Bool(true))` -> `annotations.Set(kVolatile,
true)` at the 7 call sites in
`src/s_tir/transform/{merge_shared_memory_allocations,lower_thread_allreduce}.cc`
and `src/tirx/transform/storage_rewrite.cc`. All readers use
`annotations.count(...)` (presence-only), so the value type does not
matter to consumers.

Python downstream fixes
------------------------
The original landing missed Python callers that still did `.value`
on what is now a bare int after the migration. Fixed:

  - `python/tvm/relax/transform/legalize_ops/statistical.py`:
    `_compute_shape_prod`, `_te_mean`, `_te_variance`, `_te_median`,
    `_te_std` and `_te_median` -- the `axis` parameter is now
    `list[int]`, drop the `_axis.value` / `axis[0].value` unboxing.
  - `python/tvm/relax/transform/legalize_ops/manipulate.py`: stack's
    `call.attrs.axis.value` -> `call.attrs.axis` (StackAttrs.axis is
    `Optional[int64_t]` now).

Test mirror fixes
------------------
Tests that constructed s_tir attr via `tirx.IntImm("bool", 1)` and
compared structurally now mismatch the bare `True` produced by
`@T.prim_func(s_tir=True)`. Updated 3 tests to use plain `True`:

  - `tests/python/tirx-base/test_tir_stmt_functor_substitute.py`
  - `tests/python/tirx-transform/test_tir_transform_convert_ssa.py`
  - `tests/python/tvmscript/test_tvmscript_ir_builder_tir.py` (2x)

Verified
---------
Local CPU build clean. Test sweeps green:
  - tests/python/ir/                         90 passed
  - tests/python/tirx-base/                  288 passed, 25 skipped
  - tests/python/tirx-transform/             349 passed (1 pre-existing
    LLVM-20-only failure in test_vectorize_llvm_pure_intrin_fail,
    unrelated)
  - tests/python/tvmscript/                  769 passed, 1 skipped
  - tests/python/relax/test_op_*.py          1164 passed, 5 skipped
  - tests/python/relax/test_transform_*.py   908 passed (1 pre-existing
    failure in test_transform_meta_schedule_tuning -- see note below)

Out-of-scope / not fixed (would require API design judgment)
------------------------------------------------------------
`src/s_tir/schedule/trace.cc` JSON decision deserializer produces
`Array<IntImm>` from JSON literals, but the migrated instruction
traits (e.g. `SamplePerfectTileTraits::UnpackedApplyToSchedule`) now
expect `Optional<Array<int64_t>>`. The FFI layer does not convert
`Array<IntImm>` to `Array<int64_t>` element-wise across containers.
This breaks trace replay from JSON in meta_schedule. Needs a separate
fix to either (a) convert decisions at parse time in `Trace::FromJSON`
or (b) widen the trait signature accept-types. Pre-existing in PR #19614
landing, not introduced here.

+++ src/tirx/op/runtime.cc
@@ -30,12 +30,12 @@ namespace tirx {
 TVM_REGISTER_OP("tirx.TVMBackendAnyListSetPackedArg")
     .set_num_inputs(5)
     .set_attr<TGlobalSymbol>("TGlobalSymbol", "TVMBackendAnyListSetPackedArg")
-    .set_attr<TCallEffectKind>("TCallEffectKind", Integer(CallEffectKind::kOpaque));
+    .set_attr<TCallEffectKind>("TCallEffectKind", static_cast<int64_t>(CallEffectKind::kOpaque));
 
 TVM_REGISTER_OP("tirx.TVMBackendAnyListMoveFromPackedReturn")
     .set_num_inputs(3)
     .set_attr<TGlobalSymbol>("TGlobalSymbol", "TVMBackendAnyListMoveFromPackedReturn")
-    .set_attr<TCallEffectKind>("TCallEffectKind", Integer(CallEffectKind::kOpaque));
+    .set_attr<TCallEffectKind>("TCallEffectKind", static_cast<int64_t>(CallEffectKind::kOpaque));
 
 }  // namespace tirx
 }  // namespace tvm
